Use function args, set values on inital mute, add auto mode for IR
Use Function Args:
We had some methods that were using the shared state even though they
were being passed args. Now they use the args instead. It just makes
sense!

Set Values On Initial Mute:
We have a lot of logic that depends on the global state being updated
when we are sending values (either button board to IR board or IR board
to IR transmitters). In this case, we were not updating the global
channel state on the button board when we set the initial mute, so it
thought it was on Bank 0 Channel 0 and would not respond to the button
for the first TV until it muted again (which it also only did because
  the global state did not reflect the muted state).

Auto Mode for IR
I just stole the auto increment from the button board and added it to
the IR board to test when the buttons weren't working.